backend/requirements.txt
requests
fastapi>=0.95.1
uvicorn>=0.22.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
websockets>=11.0.3
python-multipart>=0.0.6
httpx>=0.24.1
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
sqlalchemy>=2.0.0
alembic>=1.11.0 
